+++ neural.h
@@ -6,7 +6,7 @@ namespace cv {
 class Mat;
 }
 
-void initialize_samanet(int numInputLayers, float sampleRate = 30.f);
+void initialize_samanet(int numInputLayers, double sampleRate = 30.f);
 
-float run_samanet(cv::Mat &statFrame, std::vector<float> &in, float error);
+double run_samanet(cv::Mat &statFrame, std::vector<double> &in, double error);
 void save_samanet();
